Moves ELogger::verifyPath() to EPath::verify()
